`retain_tags` option for `minimap2.Mapper`.
The idea behind this change is simple. In FASTQ files generated
from `ccs` BAM files, the tags `np` and `rq` are in the FASTQ
header and give the number of passes and quality for that sequence.
We would like to retain these tags in the output SAM alignment
generated by `Mapper.map_to_sam` so that they can then be
kept in the data frame created by `Targets.parse_alignment`.

But `minimap2` does not keep these FASTQ tags in the SAM file it
creates. So this option allows `Mapper` to add the tags from
the FASTQ query file back to the created SAM alignment. This is
explained the `Mapper` docstring.

Although the idea is simple, it requires quite a bit of code in
`Mapper.map_to_sam` as the alignment has to be written to a SAM,
and then that file and the FASTQ query file jointly read to find
the tags for each read and add them into the final SAM file.