`parse_alignment_cs` returns dict of dfs
`parse_alignment_cs` returns dictionary described in docstring,
but I have not written any tests.
Currently, it builds up a single dictionary with a separate
dictionary of lists for each target. This seems to be quite fast,
but you could imagine building a separate dict of lists for each
target and then putting them all in the large final dictionary
after making dataframes might be better?