check feature/col overlap in `parse_alignment_cs`
`Targets.parse_alignment_cs` returns a data frame
with column for every feature **plus** a few columns
with fixed specified names (e.g., 'query_name'). If a
feature has the same name as one of those fixed column
names, this will lead to severe problems as two types
of information will be recorded in the same column.

Therefore, added a check to `Targets.parse_alignment_cs`
that there is not overlap in these column names and the
feature names.

Along with this change, internally got rid of the
`defaultdict` as it was no longer needed as we had to
manually enumerate the fixed column names anywhere, and
it's probably safer to use a `dict` that raises an
error if there is an attempt to access and invalid key.