fully implemented `Targets.parse_alignment`
The new `Targets.parse_alignments` is fully implemented
and tested against `Targets._parse_alignments_csv` and
for consistency in writing CSVs versus returning data frames.

It still needs more testing for correctness of output and
illustrative example.

Also, added a parameter to `Targets` explicitly permittting
the return of mutations / sequences of features with clipping;
otherwise this is disallowed as it can give confusing results.