optimize by caching results
The initial profiling indicated that the largest
cost was calling the functions in `cs_tag`. Many
of these functions were called repeatedly on the same
`cs` tag strings as we got different pieces of info
about the tags, and since some of the `cs` tags will
tend to be very similar among different alignments of
when the queries are all aligned to the same target.

It turns out that `functools.lru_cache` can be easily
used to cache results so they aren't re-computed (see
[here](https://www.cameronmacleod.com/blog/python-lru-cache)).

Added that decorator to most functions in `cs_tag`.
According to the profiling, this led to a >2-fold
speed-up in the code (compare `profile_recA.ipynb` between
the current commit and the earlier commit 31f88b892).