`ccs.Summaries` works without report, `np` tag
The original construction of `ccs.Summaries` **required**
a `ccs` report and `np` and `rq` tag in the FASTQ file. But
for data on the SRA, those tags will not be present.

Therefore, the way that `ccs.Summaries` handles this information
has been changed. The `accuracy` (name changed to this from
`quality`) is now computed directly from the Q-values, which will
always be present in a FASTQ. The number of passes is still computed
from the `np` tag, but this tag is now optional. In addition, it
is optional to provide a report.

Finally, the notebook examples have been updated to reflect this.
The notebook just showing how to use CCS summaries has been eliminated
as it is redundant with the RecA example notebook. A new test has
been created in the `./tests/` directory to confirm the absence
of `np` tags and reports is handled OK.