README better explains Python versions
In issue 58, @afrubin pointed out that the README did
not accurately described the Python version reqs (see
https://github.com/jbloomlab/alignparse/issues/58#issuecomment-563009395).

The descriptions are correct in the installation docs. Rather
than updating the README in a way that keeps the information
in two places (and so will be prone to becoming wrong again),
I've simply streamlined the README to point to the installation
docs for details.