added `select_target_names` option to `Targets`
The addition of `select_target_names` makes it possible
to create a `Targets` object for just some of the sequences
in the input `seqsfile`, which is useful if your Genbank file
contains more sequences than you want to align against.

Two cells have been added to `lasv_pilot.ipynb` that demonstrate
and teset this option.

In this process a previously silent bug in a minor check was found
on [this line](https://github.com/jbloomlab/alignparse/blob/0.1.1/alignparse/targets.py#L445),
and has now been fixed.